Reworking more tests.
