+++ src/test/java/org/idpf/epubcheck/api/Epub20CheckTest.java
@@ -20,7 +20,7 @@
  *
  */
 
-package com.adobe.epubcheck.api;
+package org.idpf.epubcheck.api;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -28,6 +28,7 @@ import java.util.List;
 
 import org.junit.Test;
 
+import com.adobe.epubcheck.api.*;
 import com.adobe.epubcheck.messages.MessageId;
 
 public class Epub20CheckTest extends AbstractEpubCheckTest
