+++ src/test/java/org/idpf/epubcheck/api/Epub30CheckExpandedTest.java
@@ -20,13 +20,14 @@
  *
  */
 
-package com.adobe.epubcheck.api;
+package org.idpf.epubcheck.api;
 
 import java.util.Collections;
 import java.util.Locale;
 
 import org.junit.Test;
 
+import com.adobe.epubcheck.api.*;
 import com.adobe.epubcheck.messages.MessageId;
 
 public class Epub30CheckExpandedTest extends AbstractEpubCheckTest
