+++ src/test/java/org/idpf/epubcheck/api/Epub30CheckTest.java
@@ -20,12 +20,13 @@
  *
  */
 
-package com.adobe.epubcheck.api;
+package org.idpf.epubcheck.api;
 
 import java.util.Collections;
 
 import org.junit.Test;
 
+import com.adobe.epubcheck.api.*;
 import com.adobe.epubcheck.messages.MessageId;
 
 public class Epub30CheckTest extends AbstractEpubCheckTest
