+++ src/test/resources/org/idpf/epubcheck/api/InputStream_Expected.txt
