+++ src/test/resources/org/idpf/epubcheck/pkg/20-invalid-mimetype/mimetype
@@ -1,0 +1,1 @@
+application/something+else
