+++ src/test/resources/org/idpf/epubcheck/pkg/ncx/20-invalid-badpath/mimetype
@@ -1,0 +1,1 @@
+application/epub+zip
