+++ src/test/resources/org/idpf/epubcheck/pkg/ncx/20-invalid-pagetarget/mimetype
@@ -1,0 +1,1 @@
+application/epub+zip
