+++ src/test/resources/org/idpf/epubcheck/pkg/opf/20-valid-uid-with-spaces/OPS/package.opf
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<package xmlns="http://www.idpf.org/2007/opf" xmlns:opf="http://www.idpf.org/2007/opf" version="2.0" unique-identifier="uid">
+  <metadata xmlns:dc="http://purl.org/dc/elements/1.1/">
+    <dc:identifier id="uid" opf:scheme="URI">urn:uuid:550e8400-e29b-41d4-a716-446667441231</dc:identifier>
+    <dc:identifier opf:scheme="ISBN"> 978-0-9852681-0-7 </dc:identifier>
+    <dc:title id="title">Minimal EPUB 2.0</dc:title>
+    <dc:language>en</dc:language>
+  </metadata>
+  <manifest>
+    <item id="content_001"  href="content_001.xhtml" media-type="application/xhtml+xml"/>
+    <item id="ncx" href="toc.ncx" media-type="application/x-dtbncx+xml" />
+  </manifest>
+  <spine toc="ncx">
+    <itemref idref="content_001" />
+  </spine>
+</package>
