+++ src/test/resources/org/idpf/epubcheck/pkg/opf/20-valid-uid-with-spaces/mimetype
@@ -1,0 +1,1 @@
+application/epub+zip
